pyproject.toml
[project]
name = "zabbixci"
description = "Source control for Zabbix templates in Git"
readme = "README.md"
dynamic = ["version"]
keywords = ["zabbix", "git", "source control", "templates"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Version Control",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking :: Monitoring",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Environment :: Console",
]
authors = [{ name = "Retigra", email = "info@retigra.nl" }]

maintainers = [
    { name = "Wouter de Bruijn", email = "wouter.debruijn@retigra.nl" },
    { name = "Raymond Kuiper", email = "raymond.kuiper@retigra.nl" },
]

dependencies = [
    "aiohttp>=3.11.11",
    "pygit2>=1.17.0",
    "regex>=2024.11.6",
    "ruamel.yaml>=0.18.10",
    "zabbix_utils>=2.0.3",
    "Wand>=0.6.13",
]
license = { file = "LICENSE.txt" }

[project.urls]
Homepage = "https://github.com/retigra/zabbixci"
Repository = "https://github.com/retigra/zabbixci"
Issues = "https://github.com/retigra/zabbixci/issues"
Changelog = "https://github.com/retigra/zabbixci/releases"
Discussions = "https://github.com/retigra/ZabbixCI/discussions"

[project.scripts]
zabbixci = "zabbixci.cli:parse_cli"

[project.optional-dependencies]
dev = ["black>=24.10.0", "isort>=5.13.2", "flake8>=7.1.1"]

[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["zabbixci", "zabbixci.*"]

[tool.setuptools_scm]
version_file = "zabbixci/_version.py"

[tool.ruff.lint]
select = ["E", "F", "B", "ERA001", "S501", "DJ", "G", "T20", "PIE", "N"]

# Ignore line-length
ignore = ["E501"]
